+++ image/resources/scripts/1_build_image.sh
@@ -17,10 +17,9 @@
 # Exit on command failure
 set -e
 
-# Set shell color variables
+# Set variables
 SHELL_YELLOW='\033[0;33m'
 SHELL_DEFAULT='\033[0m'
-
 export NEEDRESTART_MODE=a
 export NEEDRESTART_SUSPEND=1
 export DEBIAN_FRONTEND=noninteractive
@@ -29,149 +28,117 @@ export DEBIAN_PRIORITY=critical
 export QUILT_PATCHES=debian/patches
 export NAME=build EMAIL=build
 
-patches="$(pwd)/patches"
+# begin_command() formats the terminal for a command output
+begin_command() {
+    echo -e "${SHELL_YELLOW}"
+    echo "RUNNING STEP: $1..."
+    echo -e "------${SHELL_DEFAULT}"
+}
+
+# begin_command() formats the terminal after command completion
+complete_command() {
+    echo -e -n "${SHELL_YELLOW}------ "
+    echo "DONE"
+    echo -e "${SHELL_DEFAULT}"
+
+}
 
 # install_nfs_packages() installs NFS Packages
 install_nfs_packages() {
 
-    # Install Cachefilesd
-    echo "Installing cachefilesd and rpcbind..."
-    echo -e "------${SHELL_DEFAULT}"
+    begin_command "Installing rpcbind, nfs-kernel-server and cachefilesd"
     apt-get update
-    apt-get install -y rpcbind=1.2.5-8 nfs-kernel-server=1:1.3.4-2.5ubuntu3.4
+    apt-get install -y rpcbind=1.2.6-2build1 nfs-kernel-server=1:2.6.1-1ubuntu1.1 cachefilesd=0.10.10-0.2ubuntu1
     systemctl disable nfs-kernel-server
     systemctl disable nfs-idmapd.service
-    echo -e -n "${SHELL_YELLOW}------"
-    echo "DONE"
+    systemctl disable cachefilesd
+    echo "RUN=yes" >> /etc/default/cachefilesd
+    complete_command
 
 }
 
 # install_build_dependencies() installs the dependencies to required to build the kernel
 install_build_dependencies() {
 
-    echo -e "${SHELL_YELLOW}"
-    echo "Installing build dependencies..."
-    echo -e "------${SHELL_DEFAULT}"
+    begin_command "Installing build dependencies"
     apt-get update
     apt-get install -y \
         libtirpc-dev libncurses-dev flex bison openssl libssl-dev dkms \
         libelf-dev libudev-dev libpci-dev libiberty-dev autoconf dwarves \
         build-essential libevent-dev libsqlite3-dev libblkid-dev \
-        libkeyutils-dev libdevmapper-dev cdbs debhelper ubuntu-dev-tools
-    echo -e -n "${SHELL_YELLOW}------ "
-    echo "DONE"
+        libkeyutils-dev libdevmapper-dev cdbs debhelper ubuntu-dev-tools \
+        gawk llvm
+    complete_command
 
 }
 
 # download_nfs-utils() downloads version 2.5.3 of nfs-utils
 download_nfs-utils() (
 
-    echo -e "${SHELL_YELLOW}"
-    echo "Downloading nfs-utils..."
+    begin_command "Downloading nfs-utils"
     echo -e "------${SHELL_DEFAULT}"
     curl -o nfs-utils-2.5.3.tar.gz https://mirrors.edge.kernel.org/pub/linux/utils/nfs-utils/2.5.3/nfs-utils-2.5.3.tar.gz
     tar xvf nfs-utils-2.5.3.tar.gz
-    echo -e -n "${SHELL_YELLOW}------"
-    echo "DONE"
+    complete_command
 
 )
 
 # build_install_nfs-utils() builds and installs nfs-utils
 build_install_nfs-utils() (
 
-    echo -e "${SHELL_YELLOW}"
-    echo "Installing nfs-utils..."
-    echo -e "------${SHELL_DEFAULT}"
+    begin_command "Building and installing nfs-utils"
     cd nfs-utils-2.5.3
     ./configure --prefix=/usr --sysconfdir=/etc --sbindir=/sbin --disable-gss
-    make -j20
-    make install -j20
+    make -j$((`nproc`+1))
+    make install -j$((`nproc`+1))
     chmod u+w,go+r /sbin/mount.nfs
     chown nobody:nogroup /var/lib/nfs
-    echo -e -n "${SHELL_YELLOW}------"
-    echo "DONE"
-
-)
-
-# install_cachefilesd() builds and installs cachefilesd 0.10.10 from source
-install_cachefilesd() (
-    echo -e "${SHELL_YELLOW}"
-    echo "Downloading cachefilesd..."
-    echo -e "------${SHELL_DEFAULT}"
-    pull-lp-source cachefilesd 0.10.10-0.2ubuntu1
-
-    echo -e "${SHELL_YELLOW}"
-    echo "Building cachefilesd..."
-    echo -e "------${SHELL_DEFAULT}"
-
-    cd cachefilesd-0.10.10
-    quilt import "${patches}"/cachefilesd/*.patch
-    quilt push -a
-    debchange --local +knfsd "Apply knfsd patches"
-    dpkg-buildpackage -b -us -uc
-
-    echo -e "${SHELL_YELLOW}"
-    echo "Installing cachefilesd..."
-    echo -e "------${SHELL_DEFAULT}"
-
-    cd ..
-    apt-get install -y ./cachefilesd_0.10.10-0.2ubuntu1+knfsd1_amd64.deb
-    echo "RUN=yes" >> /etc/default/cachefilesd
-    systemctl disable cachefilesd
-
-    echo -e -n "${SHELL_YELLOW}------"
-    echo "DONE"
+    complete_command
 
 )
 
 # install_stackdriver_agent() installs the Cloud Ops Agent for metrics
 install_stackdriver_agent() {
 
-    echo -e "${SHELL_YELLOW}"
-    echo "Installing Cloud Ops Agent dependencies..."
-    echo -e "------${SHELL_DEFAULT}"
+    begin_command "Installing Cloud Ops Agent dependencies"
     cd ops-agent
     curl -sSO https://dl.google.com/cloudagents/add-google-cloud-ops-agent-repo.sh
-    bash add-google-cloud-ops-agent-repo.sh --also-install --version=2.11.0
+    bash add-google-cloud-ops-agent-repo.sh --also-install --version=2.22.0
     systemctl disable google-cloud-ops-agent
     cp google-cloud-ops-agent.conf /etc/logrotate.d/
     cd ..
-    echo -e -n "${SHELL_YELLOW}------ "
-    echo "DONE"
+    complete_command
 
 }
 
 # install_golang() installs golang
 install_golang() {
 
-    echo "Installing golang...."
-    echo -e "------${SHELL_DEFAULT}"
+    begin_command "Installing golang"
     curl -o go1.17.3.linux-amd64.tar.gz https://dl.google.com/go/go1.17.3.linux-amd64.tar.gz
     rm -rf /usr/local/go && tar -C /usr/local -xzf go1.17.3.linux-amd64.tar.gz
     export PATH=$PATH:/usr/local/go/bin
-    echo -e -n "${SHELL_YELLOW}------ "
-    echo "DONE"
+    complete_command
 
 }
 
 # install_knfsd_agent() installs the knfsd-agent (see https://github.com/GoogleCloudPlatform/knfsd-cache-utils/tree/main/image/knfsd-agent)
 install_knfsd_agent() (
 
-    echo "Installing knfsd-agent...."
-    echo -e "------${SHELL_DEFAULT}"
+    begin_command "Installing Knfsd agent"
     cd knfsd-agent/src
     go build -o /usr/local/bin/knfsd-agent *.go
     cd ..
     cp knfsd-logrotate.conf /etc/logrotate.d/
     cp knfsd-agent.service /etc/systemd/system/
-    echo -e -n "${SHELL_YELLOW}------ "
-    echo "DONE"
+    complete_command
 
 )
 
+# Install_knfsd_metrics_agent() installs the custom Knfsd Metrics Agent
 install_knfsd_metrics_agent() (
-    echo "Installing knfsd-metrics-agent...."
-    echo -e "------${SHELL_DEFAULT}"
+
+    begin_command "Installing knfsd-metrics-agent"
 
     cd knfsd-metrics-agent
     go build -o /usr/local/bin/knfsd-metrics-agent
@@ -180,59 +147,52 @@ install_knfsd_metrics_agent() (
     cp config/*.yaml /etc/knfsd-metrics-agent/
     cp systemd/proxy.service /etc/systemd/system/knfsd-metrics-agent.service
 
-    echo -e -n "${SHELL_YELLOW}------ "
-    echo "DONE"
+    complete_command
+
 )
 
-install_knfsd_cull() (
-    echo "Installing knfsd-cull...."
-    echo -e "------${SHELL_DEFAULT}"
-
-    cd knfsd-cull
-    go test ./...
-    go build -o /usr/local/bin/knfsd-cull
-    cp knfsd-cull.conf /etc/
-    cp knfsd-cull.service /etc/systemd/system/
-
-    echo -e -n "${SHELL_YELLOW}------ "
-    echo "DONE"
-)
-
+# install_filter_exports installs the agent that filters NFS Exports
 install_filter_exports() (
-    echo "Installing netapp-exports...."
-    echo -e "------${SHELL_DEFAULT}"
+    begin_command "Installing filter-exports"
     cd filter-exports
     go test ./...
     go build -o /usr/local/bin/filter-exports
-    echo -e -n "${SHELL_YELLOW}------ "
-    echo "DONE"
+    complete_command
 )
 
+# install_netapp_exports() installs the NetApp export detection service
 install_netapp_exports() (
-    echo "Installing netapp-exports...."
-    echo -e "------${SHELL_DEFAULT}"
+    begin_command "Installing netapp-exports"
     cd netapp-exports
     go test ./...
     go build -o /usr/local/bin/netapp-exports
     echo -e -n "${SHELL_YELLOW}------ "
-    echo "DONE"
+    complete_command
 )
 
-# install_kernel() installs the latest 5.13 kernel
-install_kernel() {
+# build_install_kernel() builds and installs the kernel with custom patches
+build_install_kernel() {
 
-    # Install the new kernel using dpkg
-    echo "Installing kernel...."
-    echo -e "------${SHELL_DEFAULT}"
-    apt-get install -y 'linux-generic-hwe-20.04=5.13.*' \
-        'linux-image-generic-hwe-20.04=5.13.*' \
-        'linux-headers-generic-hwe-20.04=5.13.*'
-
-    echo -e -n "${SHELL_YELLOW}------"
-    echo "DONE"
+    # Build and install the new kernel
+    begin_command "Building and installing kernel"
+    git clone --branch nfs-fscache-netfs --depth 1 --recurse-submodules --shallow-submodules https://github.com/benjamin-maynard/kernel.git kernel/v6.1-rc5
+    cd kernel/v6.1-rc5
+    git checkout 52acbd4584d1b83c844371e48de1a1e39d255a6d
+    cp /boot/config-`uname -r` .config
+    scripts/config --disable CONFIG_SYSTEM_REVOCATION_KEYS
+    scripts/config --disable CONFIG_SYSTEM_TRUSTED_KEYS
+    make olddefconfig
+    make clean
+    make ARCH=$(arch) -j$((`nproc`+1))
+    make ARCH=$(arch) modules_install -j$((`nproc`+1))
+    make ARCH=$(arch) install -j$((`nproc`+1))
+    cd ../../
+    rm -rf kernel/
+    complete_command
 
 }
 
+# copy_config() copies the NFS Server configuration
 copy_config() {
     chown --recursive root:root etc
     chmod --recursive 0644 etc
@@ -240,24 +200,18 @@ copy_config() {
     mkdir -p /srv/nfs
 }
 
-# Prep Server
+# Run Build
 install_nfs_packages
 install_build_dependencies
 download_nfs-utils
 build_install_nfs-utils
-install_cachefilesd
 install_stackdriver_agent
 install_golang
 install_knfsd_agent
 install_knfsd_metrics_agent
-install_knfsd_cull
 install_filter_exports
 install_netapp_exports
-
-# 5.13 HWE packages have been removed from the 20.04 distro.
-# Currently using a GCP image that comes with 5.13.
-# install_kernel
-
+build_install_kernel
 copy_config
 
 echo